+++ .openshiftio/application.yaml
@@ -101,7 +101,7 @@ objects:
       provider: fabric8
       version: "1"
       version: "1"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: my-database-secret
   stringData:
     user: luke
@@ -123,7 +123,7 @@ objects:
       app: spring-boot-hibernate-search
       provider: fabric8
       version: "1"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-hibernate-search
   spec:
     ports:
@@ -134,7 +134,7 @@ objects:
     selector:
       app: spring-boot-hibernate-search
       provider: fabric8
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
@@ -150,7 +150,7 @@ objects:
       app: spring-boot-hibernate-search
       provider: fabric8
       version: "1"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-hibernate-search
   spec:
     replicas: 1
@@ -158,7 +158,7 @@ objects:
     selector:
       app: spring-boot-hibernate-search
       provider: fabric8
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -177,7 +177,7 @@ objects:
           app: spring-boot-hibernate-search
           provider: fabric8
           version: "1"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -231,7 +231,7 @@ objects:
       app: spring-boot-hibernate-search
       provider: fabric8
       version: "1"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-hibernate-search
   spec:
     port:
