+++ .openshiftio/description.adoc
@@ -1,1 +1,1 @@
-Booster to expose an HTTP endpoint with Hibernate Search operations on fruits data.
+Example to expose an HTTP endpoint with Hibernate Search operations on fruits data.
