+++ src/main/java/dev/snowdrop/example/service/BookEnum.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 import java.time.LocalDate;
