+++ src/main/java/dev/snowdrop/example/service/BookQuery.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 public class BookQuery extends Book {
     private int maxResults = 20;
